hot/pyproject.toml
[tool.poetry]
name = "hot"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.51.2"
firecrawl-py = "^1.3.0"
flask = "^3.0.3"

[tool.poetry.group.dev.dependencies]
mypy = "^1.11.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
strict = true
